Add files via upload
Working bouncycastle code for C# is sparse and it takes a lot of time to put the small pieces together. Documentation is either non existing, hard to find or if found, reading and trying to understand may cause an inferiority complex. Anyhow, find in this repo fully working VS17 code, no fuzzing around is needed...just unzip the Bouncycastle inside the same folder and you will be able to run the code. Hope my small contribution will save some of you a couple of hours. Oh, and by the way, the BC dotnet standard 2.0 is compiling as well..just in case you need to debug your future bc based code base.